requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
scikit-learn
reportlab
fpdf2